OPTIONS channel weight chanced into float

+++ options/train_options.py
@@ -28,11 +28,11 @@ class TrainOptions(BaseOptions):
         parser.add_argument('--lr_policy', type=str, default='linear', help='learning rate policy. [linear | step | plateau | cosine]')
         parser.add_argument('--lr_decay_iters', type=int, default=50, help='multiply by a gamma every lr_decay_iters iterations')
         #
-        parser.add_argument('--weight_for_R_channel', type=int, default=0.1,
+        parser.add_argument('--weight_for_R_channel', type=float, default=0.1,
                             help='R channel is multiplied by weight before calculatin L1 loss')
-        parser.add_argument('--weight_for_G_channel', type=int, default=0.1,
+        parser.add_argument('--weight_for_G_channel', type=float, default=0.1,
                             help='R channel is multiplied by weight before calculatin L1 loss')
-        parser.add_argument('--weight_for_B_channel', type=int, default=0.1,
+        parser.add_argument('--weight_for_B_channel', type=float, default=0.1,
                             help='R channel is multiplied by weight before calculatin L1 loss')
 
         self.isTrain = True
